% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runcmd_matlab.R
\name{runcmd_matlab}
\alias{runcmd_matlab}
\title{Run MATLAB command with system command in R}
\usage{
runcmd_matlab(cmd)
}
\arguments{
\item{cmd}{A character string specifying the MATLAB command to run}
}
\value{
No return value
}
\description{
This function runs a MATLAB command with the system command in R
and saves the output to a log file.
}
\examples{
# Run a MATLAB command and save output to a log file
runcmd_matlab("disp('Hello, MATLAB!')")
}
